NPI-analysis1.txt
This R code reads in a CSV of air emission data from the NPI over the past three years. It then calculates a summary for air point emissions of Acetaldehyde in the coal mining industry. The template could be used to create a similar summary of all substances. Future work would take a functional approach to automate the work above.